+++ libseccomp/src/lib.rs
@@ -1341,6 +1341,24 @@ impl ScmpFilterContext {
         Ok(ret != 0)
     }
 
+    /// Gets the current state of the [`ScmpFilterAttr::CtlSsb`] attribute.
+    ///
+    /// This function returns `Ok(true)` if the [`ScmpFilterAttr::CtlSsb`] attribute set to on the filter being
+    /// loaded, `Ok(false)` otherwise.
+    /// The [`ScmpFilterAttr::CtlSsb`] attribute is only usable when the libseccomp API level 4 or higher
+    /// is supported.
+    ///
+    /// # Errors
+    ///
+    /// If this function is called with an invalid filter, an issue is encountered
+    /// getting the current state, or the libseccomp API level is less than 4, an error will be returned.
+    pub fn get_ctl_ssb(&self) -> Result<bool> {
+        ensure_supported_api("get_ctl_ssb", 4, ScmpVersion::from((2, 5, 0)))?;
+        let ret = self.get_filter_attr(ScmpFilterAttr::CtlSsb)?;
+
+        Ok(ret != 0)
+    }
+
     /// Sets a raw filter attribute value.
     ///
     /// The seccomp filter attributes are tunable values that affect how the library behaves
@@ -1443,6 +1461,28 @@ impl ScmpFilterContext {
     pub fn set_ctl_log(&mut self, state: bool) -> Result<()> {
         ensure_supported_api("set_ctl_log", 3, ScmpVersion::from((2, 4, 0)))?;
         self.set_filter_attr(ScmpFilterAttr::CtlLog, state.into())
+    }
+
+    /// Sets the state of the [`ScmpFilterAttr::CtlSsb`] attribute which will be applied on filter load.
+    ///
+    /// Settings this to on (`state` == `true`) disables Speculative Store Bypass mitigations for the filter.
+    /// The [`ScmpFilterAttr::CtlSsb`] attribute is only usable when the libseccomp API level 4 or higher
+    /// is supported.
+    ///
+    /// Defaults to off (`state` == `false`).
+    ///
+    /// # Arguments
+    ///
+    /// * `state` - A state flag to specify whether the [`ScmpFilterAttr::CtlSsb`] attribute should
+    /// be enabled
+    ///
+    /// # Errors
+    ///
+    /// If this function is called with an invalid filter, an issue is encountered
+    /// setting the attribute, or the libseccomp API level is less than 4, an error will be returned.
+    pub fn set_ctl_ssb(&mut self, state: bool) -> Result<()> {
+        ensure_supported_api("set_ctl_ssb", 4, ScmpVersion::from((2, 5, 0)))?;
+        self.set_filter_attr(ScmpFilterAttr::CtlSsb, state.into())
     }
 
     /// Outputs PFC(Pseudo Filter Code)-formatted, human-readable dump of a filter context's rules to a file.
@@ -2021,6 +2061,16 @@ mod tests {
         } else {
             assert!(ctx.set_ctl_log(true).is_err());
             assert!(ctx.get_ctl_log().is_err());
+        }
+
+        // Test for CtlSsb
+        if check_api(4, ScmpVersion::from((2, 5, 0))).unwrap() {
+            ctx.set_ctl_ssb(true).unwrap();
+            let ret = ctx.get_ctl_ssb().unwrap();
+            assert!(ret);
+        } else {
+            assert!(ctx.set_ctl_ssb(true).is_err());
+            assert!(ctx.get_ctl_ssb().is_err());
         }
     }